feat: Initialize TaskManager with a database connection
This commit initializes the TaskManager struct with a database connection. It uses the rusqlite crate to establish a connection to the "appdata.db" database file. If the connection is successful, the tasks table is created if it doesn't exist. Otherwise, an error is returned.

The purpose of this change is to integrate the database functionality into the TaskManager, allowing tasks to be stored and retrieved from the database.